requirements.txt
pandas==1.4.2
Pillow==9.1.1
streamlit==1.9.0
